revert back popularity in sequelize to double  Changes to be committed: 	deleted:    server/movies/migrations/20200216000352-change-popularity-type-to-integer.js 	modified:   server/movies/models/movies.js 	modified:   server/movies/seeders/20200213101242-movies-seeder.js 	modified:   server/orchestrator/index.js 	modified:   server/orchestrator/schemas/externalMovies.js 	modified:   server/orchestrator/schemas/movies.js 	modified:   server/orchestrator/schemas/tv.js

+++ server/movies/models/movies.js
@@ -15,7 +15,7 @@ module.exports = (sequelize, DataTypes) => {
       title: { type: DataTypes.STRING },
       overview: { type: DataTypes.STRING },
       posterPath: { type: DataTypes.STRING },
-      popularity: { type: DataTypes.INTEGER },
+      popularity: { type: DataTypes.DOUBLE },
       tags: { type: DataTypes.ARRAY(DataTypes.STRING) }
     },
     {}
